[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 1112 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -1,9 +1,9 @@
 #include<stdio.h>
 
 int main(void){
-  int a, b;
-  scanf("%d %d", &a, &b);
-  printf("%d\n", (a+b)*(a-b));
+  long long a, b;
+  scanf("%lld %lld", &a, &b);
+  printf("%lld\n", (a+b)*(a-b));
 
   return 0;
 }
